refactor: fix lint and knex logs errors

+++ src/infrastructure/database/migrations/20210814191951_table_projects.js
@@ -1,12 +1,18 @@
 const { tables } = require("../../../config/database");
 
 exports.up = function (knex) {
-  return knex.schema.createTableIfNotExists(tables.projects, (table) => {
-    table.uuid("id").primary();
-    table.string("name", 350).notNullable();
-    table.uuid("user_id").references(`${tables.users}.id`).onDelete("CASCADE").notNullable();
+  return knex.schema.hasTable(tables.projects).then(function (exists) {
+    if (!exists) {
+      return knex.schema
+        .createTable(tables.projects, function (table) {
+          table.uuid("id").primary();
+          table.string("name", 350).notNullable();
+          table.uuid("user_id").references(`${tables.users}.id`).onDelete("CASCADE").notNullable();
 
-    table.timestamps(true, true);
+          table.timestamps(true, true);
+        })
+        .then();
+    }
   });
 };
